12152 libc: cast to pointer from integer of different size
../i386/crt/_rtld.c: In function '__rtld':
../i386/crt/_rtld.c:105:21: warning: cast to pointer from integer of different size [-Wint-to-pointer-cast]
  105 |    program_name = *((char **)ebp->eb_un.eb_ptr);
      |                     ^
../i386/crt/_rtld.c:108:14: warning: cast to pointer from integer of different size [-Wint-to-pointer-cast]
  108 |    for (ap = (auxv_t *)ebp->eb_un.eb_ptr;
      |              ^
../i386/crt/_rtld.c:139:9: warning: cast to pointer from integer of different size [-Wint-to-pointer-cast]
  139 |  ehdr = (Elf32_Ehdr *)MMAP(0, sb.st_size, PROT_READ | PROT_EXEC,
      |         ^
../i386/crt/_rtld.c:197:10: warning: cast to pointer from integer of different size [-Wint-to-pointer-cast]
  197 |  maddr = (caddr_t)MMAP(0, mlen, PROT_READ | PROT_EXEC,
      |          ^
../i386/crt/_rtld.c:50:24: warning: cast from pointer to integer of different size [-Wpointer-to-int-cast]
   50 | #define ROUND(x, a)  (((int)(x) + ((int)(a) - 1)) & \
      |                        ^
../i386/crt/_rtld.c:201:19: note: in expansion of macro 'ROUND'
  201 |  faddr = (caddr_t)ROUND(maddr, fph->p_align);
      |                   ^~~~~
../i386/crt/_rtld.c:201:10: warning: cast to pointer from integer of different size [-Wint-to-pointer-cast]
  201 |  faddr = (caddr_t)ROUND(maddr, fph->p_align);
      |          ^
../i386/crt/_rtld.c:211:11: warning: cast to pointer from integer of different size [-Wint-to-pointer-cast]
  211 |   maddr = (caddr_t)MMAP(0, mlen, PROT_READ | PROT_EXEC,
      |           ^
../i386/crt/_rtld.c:50:24: warning: cast from pointer to integer of different size [-Wpointer-to-int-cast]
   50 | #define ROUND(x, a)  (((int)(x) + ((int)(a) - 1)) & \
      |                        ^
../i386/crt/_rtld.c:215:20: note: in expansion of macro 'ROUND'
  215 |   faddr = (caddr_t)ROUND(maddr, fph->p_align);
      |                    ^~~~~
../i386/crt/_rtld.c:215:11: warning: cast to pointer from integer of different size [-Wint-to-pointer-cast]
  215 |   faddr = (caddr_t)ROUND(maddr, fph->p_align);
      |           ^
../i386/crt/_rtld.c:49:23: warning: cast from pointer to integer of different size [-Wpointer-to-int-cast]
   49 | #define ALIGN(x, a)  ((int)(x) & ~((int)(a) - 1))
      |                       ^
../i386/crt/_rtld.c:242:19: note: in expansion of macro 'ALIGN'
  242 |   addr = (caddr_t)ALIGN(faddr + pptr->p_vaddr, page_size);
      |                   ^~~~~
../i386/crt/_rtld.c:242:10: warning: cast to pointer from integer of different size [-Wint-to-pointer-cast]
  242 |   addr = (caddr_t)ALIGN(faddr + pptr->p_vaddr, page_size);
      |          ^
../i386/crt/_rtld.c:250:28: warning: cast from pointer to integer of different size [-Wpointer-to-int-cast]
  250 |    (ebp++)->eb_un.eb_ptr = (Elf32_Addr)addr;
      |                            ^
../i386/crt/_rtld.c:273:7: warning: cast to pointer from integer of different size [-Wint-to-pointer-cast]
  273 |   if ((caddr_t)MMAP((caddr_t)addr, flen, i,
      |       ^
../i386/crt/_rtld.c:284:11: warning: cast from pointer to integer of different size [-Wpointer-to-int-cast]
  284 |    foff = (int)faddr + pptr->p_vaddr + pptr->p_filesz;
      |           ^
../i386/crt/_rtld.c:285:12: warning: cast to pointer from integer of different size [-Wint-to-pointer-cast]
  285 |    zaddr = (caddr_t)ROUND(foff, page_size);
      |            ^
../i386/crt/_rtld.c:286:20: warning: cast from pointer to integer of different size [-Wpointer-to-int-cast]
  286 |    for (j = 0; j < (int)(zaddr - foff); j++)
      |                    ^
../i386/crt/_rtld.c:287:7: warning: cast to pointer from integer of different size [-Wint-to-pointer-cast]
  287 |     *((char *)foff + j) = 0;
      |       ^
../i386/crt/_rtld.c:295:9: warning: cast to pointer from integer of different size [-Wint-to-pointer-cast]
  295 |     if ((caddr_t)MMAP((caddr_t)zaddr, j, i,
      |         ^

+++ usr/src/lib/libc/i386/crt/_rtld.c
@@ -102,10 +102,10 @@ __rtld(Elf32_Boot *ebp, const char *strings[], int (*funcs[])())
 	while (ebp->eb_tag != 0) {
 		switch (ebp->eb_tag) {
 		case EB_ARGV:
-			program_name = *((char **)ebp->eb_un.eb_ptr);
+			program_name = *((char **)(uintptr_t)ebp->eb_un.eb_ptr);
 			break;
 		case EB_AUXV:
-			for (ap = (auxv_t *)ebp->eb_un.eb_ptr;
+			for (ap = (auxv_t *)(uintptr_t)ebp->eb_un.eb_ptr;
 			    ap->a_type != AT_NULL; ap++)
 				if (ap->a_type == AT_PAGESZ) {
 					page_size = ap->a_un.a_val;
@@ -136,8 +136,8 @@ __rtld(Elf32_Boot *ebp, const char *strings[], int (*funcs[])())
 		PANIC(program_name);
 	if (FSTATAT(ldfd, NULL, &sb, 0) == -1)
 		PANIC(program_name);
-	ehdr = (Elf32_Ehdr *)MMAP(0, sb.st_size, PROT_READ | PROT_EXEC,
-	    MAP_SHARED, ldfd, 0);
+	ehdr = (Elf32_Ehdr *)(uintptr_t)MMAP(0, sb.st_size,
+	    PROT_READ | PROT_EXEC, MAP_SHARED, ldfd, 0);
 	if (ehdr == (Elf32_Ehdr *)-1)
 		PANIC(program_name);
 
@@ -194,11 +194,11 @@ __rtld(Elf32_Boot *ebp, const char *strings[], int (*funcs[])())
 	 */
 	mlen = ROUND((lph->p_vaddr + lph->p_memsz) -
 	    ALIGN(fph->p_vaddr, page_size), page_size);
-	maddr = (caddr_t)MMAP(0, mlen, PROT_READ | PROT_EXEC,
+	maddr = (caddr_t)(uintptr_t)MMAP(0, mlen, PROT_READ | PROT_EXEC,
 	    MAP_SHARED, ldfd, 0);
 	if (maddr == (caddr_t)-1)
 		PANIC(program_name);
-	faddr = (caddr_t)ROUND(maddr, fph->p_align);
+	faddr = (caddr_t)(uintptr_t)ROUND((uintptr_t)maddr, fph->p_align);
 
 	/*
 	 * Check to see whether alignment skew was really needed.
@@ -208,11 +208,12 @@ __rtld(Elf32_Boot *ebp, const char *strings[], int (*funcs[])())
 		mlen = ROUND((lph->p_vaddr + lph->p_memsz) -
 		    ALIGN(fph->p_vaddr, fph->p_align) + fph->p_align,
 		    page_size);
-		maddr = (caddr_t)MMAP(0, mlen, PROT_READ | PROT_EXEC,
+		maddr = (caddr_t)(uintptr_t)MMAP(0, mlen, PROT_READ | PROT_EXEC,
 		    MAP_SHARED, ldfd, 0);
 		if (maddr == (caddr_t)-1)
 			PANIC(program_name);
-		faddr = (caddr_t)ROUND(maddr, fph->p_align);
+		faddr = (caddr_t)(uintptr_t)
+		    ROUND((uintptr_t)maddr, fph->p_align);
 	}
 
 	/*
@@ -239,7 +240,8 @@ __rtld(Elf32_Boot *ebp, const char *strings[], int (*funcs[])())
 		/*
 		 * Set address of this segment relative to our base.
 		 */
-		addr = (caddr_t)ALIGN(faddr + pptr->p_vaddr, page_size);
+		addr = (caddr_t)(uintptr_t)
+		    ALIGN((uintptr_t)faddr + pptr->p_vaddr, page_size);
 
 		/*
 		 * If this is the first program header, record our base
@@ -247,7 +249,7 @@ __rtld(Elf32_Boot *ebp, const char *strings[], int (*funcs[])())
 		 */
 		if (pptr == phdr) {
 			ebp->eb_tag = EB_LDSO_BASE;
-			(ebp++)->eb_un.eb_ptr = (Elf32_Addr)addr;
+			(ebp++)->eb_un.eb_ptr = (Elf32_Addr)(uintptr_t)addr;
 		}
 
 		/*
@@ -270,7 +272,7 @@ __rtld(Elf32_Boot *ebp, const char *strings[], int (*funcs[])())
 			i |= PROT_WRITE;
 		if (pptr->p_flags & PF_X)
 			i |= PROT_EXEC;
-		if ((caddr_t)MMAP((caddr_t)addr, flen, i,
+		if ((caddr_t)(uintptr_t)MMAP((caddr_t)addr, flen, i,
 		    MAP_FIXED | MAP_PRIVATE, ldfd, foff) == (caddr_t)-1)
 			PANIC(program_name);
 
@@ -281,10 +283,11 @@ __rtld(Elf32_Boot *ebp, const char *strings[], int (*funcs[])())
 		 * map some more space from /dev/zero.
 		 */
 		if (pptr->p_memsz > pptr->p_filesz) {
-			foff = (int)faddr + pptr->p_vaddr + pptr->p_filesz;
-			zaddr = (caddr_t)ROUND(foff, page_size);
-			for (j = 0; j < (int)(zaddr - foff); j++)
-				*((char *)foff + j) = 0;
+			foff = (uintptr_t)faddr + pptr->p_vaddr +
+			    pptr->p_filesz;
+			zaddr = (caddr_t)(uintptr_t)ROUND(foff, page_size);
+			for (j = 0; j < (uintptr_t)(zaddr - foff); j++)
+				*((char *)(uintptr_t)foff + j) = 0;
 			j = (faddr + pptr->p_vaddr + pptr->p_memsz) - zaddr;
 			if (j > 0) {
 				if (dzfd == 0) {
@@ -292,8 +295,8 @@ __rtld(Elf32_Boot *ebp, const char *strings[], int (*funcs[])())
 					if (dzfd == -1)
 						PANIC(program_name);
 				}
-				if ((caddr_t)MMAP((caddr_t)zaddr, j, i,
-				    MAP_FIXED | MAP_PRIVATE, dzfd,
+				if ((caddr_t)(uintptr_t)MMAP((caddr_t)zaddr,
+				    j, i, MAP_FIXED | MAP_PRIVATE, dzfd,
 				    0) == (caddr_t)-1)
 					PANIC(program_name);
 			}
